refactor context loading
