+++ python/citation_vim/context.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+class Context(object):
+    'empty context class'
